requirements.txt
pycountry
psycopg2-binary
thefuzz[speedup]
python-dotenv
